split author information to 3 filters
